main_server/requirements.txt
fastapi
uvicorn
httpx
opencv-python
numpy
psycopg2-binary